DF-219 - first draft, move AvamCreateJobAdvertisementDto to infrasctructure-messagebroker-shared
